refactor(reference): promote the content-identity fingerprint into a shipping module
Move fnv1a_hex/reference_identity_from_manifest/inject_content_stamps and
CONTENT_STAMPED_ARTIFACTS verbatim from the mutalyzer test into
src/prepare/identity.rs so ferro prepare and the loader can share one
definition (#1001). No algorithm change: the identity is byte-identical.
Adds an invariance test that the schema version and reference_identity
field are excluded from the hash.
